Implemented single responsibility, open/close and liskov principle.

+++ app/src/main/java/com/alexcfa/yourweather/data/RemoteResult.kt
@@ -103,7 +103,7 @@ data class Current(
     @SerialName("pressure")
     val pressure: Int,
     @SerialName("precip")
-    val precip: Int,
+    val precip: Double,
     @SerialName("humidity")
     val humidity: Int,
     @SerialName("cloudcover")
